--- Source/Tests/Paths/Setup/PathsCollection.cs
@@ -1,6 +1,0 @@
-namespace Bloodmasters.Tests.Paths.Setup;
-
-[CollectionDefinition(TestCollections.Paths)]
-public class PathsCollection : ICollectionFixture<PathsFixture>
-{
-}
