+++ Source/Tests/Paths/TempDirPathTests.cs
@@ -1,6 +1,5 @@
 namespace Bloodmasters.Tests.Paths;
 
-[Collection(TestCollections.Paths)]
 public sealed class TempDirPathTests
 {
     [Fact(DisplayName = "Temporary directory should contain suffix 'Bloodmasters'")]
