--- Source/Tests/TestCollections.cs
@@ -1,6 +1,0 @@
-namespace Bloodmasters.Tests;
-
-public static class TestCollections
-{
-    public const string Paths = "Paths collection";
-}
